requirements.txt
Flask
openai
twilio
elevenlabs
python-dotenv
requests 